lib/iocache: Fix signedness error in iocache_read()
When iocache_read() returned -1 earlier, it would decrement the
ioc->ioc_buflen variable by 1, causing us to potentially read into
uninitialized data and missing one byte of the message. This was
caused by a flawed comparison of the (unsigned) size_t value we
got from the read against 0 to see if we had any errors. If not,
we'd add the return result to our ioc_buflen variable, causing it
to result in an offset of -1 when we next read data into the buffer.

+++ lib/iocache.c
@@ -132,7 +132,7 @@ iocache *iocache_create(unsigned long size)
 
 int iocache_read(iocache *ioc, int fd)
 {
-	size_t to_read, bytes_read;
+	int to_read, bytes_read;
 
 	if (!ioc || !ioc->ioc_buf || fd < 0)
 		return -1;
